ml-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
supabase==2.9.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
xgboost==2.1.1
shap==0.46.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
